feat: finish lc selector
